Add distillery
We'll be using distillery for releases. It's been added to our
dependencies, and `mix release.init` was run. A few of the resulting
files were modified/removed:

- `rel/plugins/` was removed
- `rel/vm.args` was removed
- `rel/config.exs` was modified to remove the reference to `vm.args`

The plugins directory isn't necessary because we won't be using any,
and `vm.args` is generated for us by Gigalixir when we deploy releases.