Add continuous deployment to Travis configuration
This requires a bit more work than you'd think. We first need to first
add gigalixir as a remote to our CI environment, which accesses keys
that need to be defined in Travis.

Travis also exposes the branch name in a bit of an indirect way. The
short version is that we need to use `TRAVIS_PULL_REQUEST_BRANCH` by
default, and fall back to `TRAVIS_BRANCH` if it's not defined. See here
for more details:

https://github.com/travis-ci/travis-ci/issues/6652

Lastly we run our tests as usual, immediately followed by a deploy if
we have determined that we're on `master`.