Add excoveralls
This adds excoveralls to our dependencies for `:test` only, as well as
configuring the mix tasks to always use the `:test` environment.

We also have a simple excoveralls configuration. Currently we are
requiring 80% test coverage, but this number may change. We're also
ignoring the `test` directory.

The .gitignore has been updated to exclude coverage files as well.