Use `System.get_env` for configuration
This seems to be a touchy subject in general, but I'm finding more and
more examples that prefer the lower-level `System.get_env` instead of
the somewhat-hacky `"${}"` syntax. This updates our prod config so that
we aren't using it anyumore.

Note that we're failing on a few keys here if nothing is defined to
protect us from starting the application in an unsafe state.